+++ owntracks-android-2.5/project/app/src/main/java/org/owntracks/android/net/WifiInfoProvider.kt
@@ -11,7 +11,6 @@ import android.os.Build
 import dagger.hilt.android.qualifiers.ApplicationContext
 import javax.inject.Inject
 import javax.inject.Singleton
-import org.owntracks.android.model.messages.MessageStatus
 
 @Singleton
 class WifiInfoProvider @Inject constructor(@ApplicationContext context: Context) {
@@ -63,11 +62,7 @@ class WifiInfoProvider @Inject constructor(@ApplicationContext context: Context)
   fun isConnected(): Boolean = getBSSID() != null
 
   fun isWiFiEnabled(): Int =
-      if (wifiManager.isWifiEnabled) {
-        MessageStatus.STATUS_WIFI_ENABLED
-      } else {
-        MessageStatus.STATUS_WIFI_DISABLED
-      }
+      if (wifiManager.isWifiEnabled) { 1 } else { 0 }
 }
 
 fun WifiInfo.getUnquotedSSID(): String = this.ssid.replace(Regex("^\"(.*)\"$"), "$1")