+++ owntracks-android-2.5/project/app/src/main/java/org/owntracks/android/services/LocationProcessor.kt
@@ -12,7 +12,6 @@ import org.owntracks.android.data.waypoints.WaypointsRepo
 import org.owntracks.android.di.ApplicationScope
 import org.owntracks.android.di.CoroutineScopes
 import org.owntracks.android.location.geofencing.Geofence
-import org.owntracks.android.model.messages.AddMessageStatus
 import org.owntracks.android.model.messages.MessageLocation
 import org.owntracks.android.model.messages.MessageLocation.Companion.fromLocation
 import org.owntracks.android.model.messages.MessageLocation.Companion.fromLocationAndWifiInfo
@@ -20,12 +19,13 @@ import org.owntracks.android.model.messages.MessageStatus
 import org.owntracks.android.model.messages.MessageTransition
 import org.owntracks.android.model.messages.MessageWaypoint
 import org.owntracks.android.model.messages.MessageWaypoints
+import org.owntracks.android.model.messages.addMessageStatus
 import org.owntracks.android.net.WifiInfoProvider
 import org.owntracks.android.preferences.Preferences
 import org.owntracks.android.preferences.types.MonitoringMode
 import org.owntracks.android.support.DeviceMetricsProvider
 import org.owntracks.android.support.MessageWaypointCollection
-import org.owntracks.android.support.SimpleIdlingResource
+import org.owntracks.android.test.SimpleIdlingResource
 import timber.log.Timber
 import java.time.Instant
 import java.util.concurrent.TimeUnit
@@ -128,6 +128,7 @@ constructor(
               address = lastAddress
             }
     Timber.v("Actually publishing location $location triggered by $trigger as message=$message")
+
     messageProcessor.queueMessageForSending(message)
     if (responseMessageTypes.contains(trigger)) {
       publishResponseMessageIdlingResource.setIdleState(true)
@@ -256,17 +257,16 @@ constructor(
     publishResponseMessageIdlingResource.setIdleState(true)
   }
 
-  fun publishStatusMessage() {
+  suspend fun publishStatusMessage() {
     messageProcessor.queueMessageForSending(
         MessageStatus().apply {
-          android =
-              AddMessageStatus().apply {
-                wifistate = wifiInfoProvider.isWiFiEnabled()
-                powerSave = deviceMetricsProvider.powerSave
-                batteryOptimizations = deviceMetricsProvider.batteryOptimizations
-                appHibernation = deviceMetricsProvider.appHibernation
-                locationPermission = deviceMetricsProvider.locationPermission
-              }
+          android = addMessageStatus().apply {
+            wifistate = wifiInfoProvider.isWiFiEnabled();
+            powerSave = deviceMetricsProvider.powerSave;
+            batteryOptimizations = deviceMetricsProvider.batteryOptimizations;
+            appHibernation = deviceMetricsProvider.appHibernation;
+            locationPermission = deviceMetricsProvider.locationPermission;
+          }
         })
     publishResponseMessageIdlingResource.setIdleState(true)
   }