+++ owntracks-android-2.5/project/app/src/main/java/org/owntracks/android/ui/map/MapViewModel.kt
@@ -177,7 +177,6 @@ constructor(
       currentLocation.value?.run {
         Timber.d("Sending current location from user request: $this")
         locationProcessor.onLocationChanged(this, MessageLocation.ReportType.USER)
-        locationProcessor.publishStatusMessage()
       }
     }
   }