+++ owntracks-android-2.5/project/app/src/test/java/org/owntracks/android/preferences/PreferencesGettersAndSetters.kt
@@ -12,6 +12,7 @@ import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
 import org.mockito.kotlin.doReturn
 import org.mockito.kotlin.mock
+import org.owntracks.android.location.LocatorPriority
 import org.owntracks.android.model.messages.MessageConfiguration
 import org.owntracks.android.preferences.types.ConnectionMode
 import org.owntracks.android.preferences.types.MonitoringMode
@@ -19,7 +20,7 @@ import org.owntracks.android.preferences.types.MqttProtocolLevel
 import org.owntracks.android.preferences.types.MqttQos
 import org.owntracks.android.preferences.types.ReverseGeocodeProvider
 import org.owntracks.android.preferences.types.StringMaxTwoAlphaNumericChars
-import org.owntracks.android.support.SimpleIdlingResource
+import org.owntracks.android.test.SimpleIdlingResource
 
 @RunWith(Parameterized::class)
 class PreferencesGettersAndSetters(private val parameter: Parameter) {
@@ -128,6 +129,12 @@ class PreferencesGettersAndSetters(private val parameter: Parameter) {
               Parameter("keepalive", -1, Int::class, false, preferenceValueExpected = 0),
               Parameter("locatorDisplacement", 1690, Int::class, false),
               Parameter("locatorInterval", 1000, Int::class, false),
+              Parameter(
+                  "locatorPriority",
+                  LocatorPriority.HighAccuracy,
+                  LocatorPriority::class,
+                  false,
+                  preferenceValueInConfiguration = "HighAccuracy"),
               Parameter(
                   "mode",
                   ConnectionMode.HTTP,