+++ owntracks-android-2.5/project/app/src/test/java/org/owntracks/android/support/ParserTest.kt
@@ -19,7 +19,6 @@ import org.mockito.kotlin.mock
 import org.owntracks.android.location.geofencing.Geofence
 import org.owntracks.android.model.BatteryStatus
 import org.owntracks.android.model.CommandAction
-import org.owntracks.android.model.messages.AddMessageStatus
 import org.owntracks.android.model.messages.Clock
 import org.owntracks.android.model.messages.MessageCard
 import org.owntracks.android.model.messages.MessageClear
@@ -31,6 +30,7 @@ import org.owntracks.android.model.messages.MessageStatus
 import org.owntracks.android.model.messages.MessageTransition
 import org.owntracks.android.model.messages.MessageUnknown
 import org.owntracks.android.model.messages.MessageWaypoint
+import org.owntracks.android.model.messages.addMessageStatus
 import org.owntracks.android.preferences.Preferences
 import org.owntracks.android.preferences.types.MonitoringMode
 import org.owntracks.android.preferences.types.MqttQos
@@ -744,14 +744,13 @@ class ParserTest {
     val parser = Parser(encryptionProvider)
     val message =
         MessageStatus().apply {
-          android =
-              AddMessageStatus().apply {
-                wifistate = 1
-                powerSave = 1
-                batteryOptimizations = 1
-                appHibernation = 1
-                locationPermission = -3
-              }
+          android = addMessageStatus().apply {
+            wifistate = 1
+            powerSave = 1
+            batteryOptimizations = 1
+            appHibernation = 1
+            locationPermission = -3
+          }
         }
     val serialized = message.toJson(parser)
     val jsonNode = objectMapper.readTree(serialized)
@@ -807,6 +806,27 @@ class ParserTest {
     assertTrue(messageCard.isValidMessage())
     assertEquals("MyName!", messageCard.name)
     assertEquals("owntracks/user/device", messageCard.getContactId())
+  }
+
+  @Test
+  fun `Parser can deserialize a MessageCard with a trackerId field`() {
+    val parser = Parser(encryptionProvider)
+    // language=JSON
+    val input =
+        """
+          {
+          "_type": "card",
+          "tid": "overridden-topic",
+          "name": "MyName!"
+          }
+        """
+            .trimIndent()
+    val messageBase = parser.fromJson(input)
+    assertEquals(MessageCard::class.java, messageBase.javaClass)
+    val messageCard = messageBase as MessageCard
+    assertTrue(messageCard.isValidMessage())
+    assertEquals("MyName!", messageCard.name)
+    assertEquals("overridden-topic", messageCard.trackerId)
   }
   // endregion